+++ .github/workflows/test.yml
@@ -27,9 +27,8 @@ jobs:
   
     - name: Run mypy
       run: |
-        python -m poetry run mypy .
+        python -m poetry run mypy texlive
 
     - name: Run flake8
       run: |
-        python -m poetry run flake8 .
-    
+        python -m poetry run flake8 texlive
